refactor(1013): extract OrgConfigMigrationManager to src/org/ (SC-001, position 5)
Cat B fresh-extraction P5. Verbatim relocation of the 678-line
OrgConfigMigrationManager class from MistHelper.py to
src/org/org_config_migration_manager.py. Menus 176 and 177 continue to
reference the class name directly via a top-of-file import; no logic
changes.

- MistHelper.py: 20740 -> 20068 lines; FR-007 breadcrumb left at the
  deletion site.
- New module: 10.00/10 pylint, black + ruff clean.
- --test exit 0, 66/66 successful operations, no regressions.
- refactor_candidates.md regenerated (FR-014).

Refs #1013 (FR-002, FR-007, FR-013, FR-014, SC-001, SC-005, SC-006).
